Fixed metatron.sh script upstart mode
The upstart mode couldn't be used because there was no logic to extract
the flag from the arguments passed to the script.

The upstart mode starts the application in foreground mode which is
useful if you use a service manager which manages the lifecycle of the
process. I renamed the upstart mode to foreground since it basically
just starts the application in the foreground. This mode is also used by
other service managers like systemd or Solaris / Illumos SMF.

I have also removed the piping to log file from the forgeround mode.
When using systems like systemd, SMF, etc. these manage the output of
the processes they manage. For example in case of systemd by displaying
it via journalctl and SMF does this similar.